add `repo` and `ref` to workflow inputs
when the reusable workflows are called from library repos like `rmm`, they check out the correct source code.

however, when the reusable workflows are called from the nightly "pipeline of pipelines" repo, they attempt to check out the repository from which that nightly workflow lives.

this change should fix that.