Remove `merge-gatekeeper`
This PR removes `merge-gatekeeper`. We have to remove it due to some rate-limiting errors we were getting on high-volume repositories like `cudf`.

According to the GH docs below, the `GITHUB_TOKEN` provided by GitHub Actions is limited to 1000 requests per repository per hour. This limit is not enough for repos like `cudf`.

https://docs.github.com/en/rest/overview/resources-in-the-rest-api?apiVersion=2022-11-28#requests-from-github-actions