Add feature to skip uploading `conda` packages
This PR adds a new input, `skip_upload_pkgs`, to the `conda-upload-packages.yaml` workflow.

The value of this input is passed as an environment variable to the `rapids-upload-to-anaconda` script.

With the PR below merged, this will allow repositories to specify a space-delimited list of `conda` package names that should not be uploaded to Anaconda.org.

- https://github.com/rapidsai/gha-tools/pull/32

It can be used like this:

```yaml
upload-conda:
    uses: rapidsai/shared-action-workflows/.github/workflows/conda-upload-packages.yaml@main
    with:
      skip_upload_pkgs: pkg1 pkg2 pkg3

upload-conda:
    uses: rapidsai/shared-action-workflows/.github/workflows/conda-upload-packages.yaml@main
    with:
      skip_upload_pkgs: >-
        pkg1
        pkg2
        pkg3
```