Update `a100` jobs to use driver `525`
I am in the middle of updating our two `arm` nodes to driver `525`.

One of them has been updated, so we should update our current `a100` jobs to use this new driver so that I can update the other one.

We can leave the `CUDA_VER` for these `a100` jobs as `11.8` on `branch-23.04` and then update them to `12.0` on branch `cuda-120`.