Update `drivers` for `branch-23.02`
This PR updates the `520` drivers used for testing in `branch-23.02` to `525`.

This is necessary since there is a hotfix for `23.02` that needs to go out and we no longer have driver `520` machines.

In the future, we'll stop using driver-version-specific label names in favor of labels like `driver-latest` and `driver-earliest`.

This workaround is fine until then.