`wheels-manylinux-test.yml` Fix (#67)
#63 included a bug that caused the wheel unit and smoke test commands to
not be properly escaped.

The fix was to inject those values as `bash` variables rather than with
a GitHub Action's expression.

This required updating the outer quotations for the `MATRICES`
environment variable in `.github/workflows/wheels-manylinux-test.yml` to
use `"` instead of `'`.

For consistency, I've updated the other matrix environment variables to
match this style.

Additionally, I've switched the final matrix output to use `tee
--append` instead of `>>` so that the matrix value is printed to the
GitHub Actions console. This is useful for inspecting the final computed
matrices.