Use new manylinux images and avoid docker in docker (#71)
Updates the wheel building workflows to use the new images introduced in
https://github.com/rapidsai/cibuildwheel-imgs/pull/13. These images
subsume both the old manylinux and cibuildwheel containers, allowing
both pure and manylinux wheels to be built in the same environment.
Since all the necessary tools are present in the same place, there is
also no longer a need to use the old Docker in Docker approach. Instead,
we eschew usage of cibuildwheel in favor of invoking `python -m pip
wheel` ourselves. We were already only making partial use of
cibuildwheel since 1) we can't leverage cibuildwheel's test infra
because we need to run tests on different runners (with GPUs) than
builds (without GPUs); and 2) we need to orchestrate builds across
different architectures, which cibuildwheel does not handle
intrinsically (although it does support different architectures). The
new approach should also make it much easier to enable local wheel
testing, although there remain a couple of minor issues in that vein
that can be dealt with in a follow-up.