Add `RAPIDS_AUX_SECRETS_1` to select workflows
There is a need to inject a secret into these shared-workflows by one of our repositories.

This PR adds an optional secret, `RAPIDS_AUX_SECRET_1`, to the necessary shared workflows.

I intentionally didn't add this secret to the other workflows, because its use is strongly discouraged. Therefore, we should only add it to workflows as needed.

This secret can be used by calling workflows like this:

```yaml
jobs:
  wheels-test:
    uses: rapidsai/shared-workflows/.github/workflows/wheels-test.yaml@branch-24.06
    secrets:
      RAPIDS_AUX_SECRET_1: ${{ secrets.REPO_SPECIFIC_SECRET }}
```